v1.5.1: 第五轮C/D批 - 行为一致(习惯非计划日/WebDAV消息/首页tab/深链/打卡互斥/启动就绪) + 加固清理(通知contentIntent与ID位段/时区重排/迁移测试到12/重复规则兜底/死代码)

+++ shared/ui/src/commonMain/kotlin/com/enlpot/daydo/shared/ui/habit/ui/sections/HabitsList.kt
@@ -56,7 +56,7 @@ fun HabitsList(
     state: HabitState,
     lazyListState: LazyListState,
     onAction: (HabitsAction) -> Unit,
-    onNavigateToAnalytics: () -> Unit,
+    onNavigateToAnalytics: (Long) -> Unit,
     modifier: Modifier = Modifier,
 ) {
     val windowSizeClass = LocalWindowSizeClass.current
